rename folder and utils
