+++ frontend/src/components/assessments/student-experience/question-elements/Image.tsx
