+++ frontend/src/components/assessments/student-experience/question-elements/QuestionText.tsx
